Move reasoning-specific code into its own module
